First "working" version
This version can launch a bare ubuntu desktop session from a
console just by running

    snap run ubuntu-desktop-session

This session still doesn't have sound, can't launch programs...
is extremely limited, but it's a beginning.

It works on a modified core24-desktop with all the GDM and Gnome
packages removed (basically the 002.1-add-gdm.chroot file is
removed), and still uses it as base.

The next step is to use core24 as base. For that, we need to
find which packages install gnome-initial-setup that are
required for DRI/DRM, because without it (installed with the
002.4-configure-system-setup-tool.chroot), gnome shell returns
an error when trying to create the graphical console.